Adding a routing to host images in the API
Added a /images/videoId route to be able to host the video images in the API server.
Added the images to the structure folder
Deleted the videos.json from the data folder, since this file won't be used anymore.
Made the needed modifications to remove any usage of the videos.json file.